% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/corpus.r
\name{corpora.list.compare}
\alias{corpora.list.compare}
\title{The corpora.compare function for a list of dtm's}
\usage{
corpora.list.compare(x, subcorpus = NULL, method = "each_to_all",
  return.df = F, chain.lag = 1, ...)
}
\arguments{
\item{x}{either a list of document term matrices, or a single dtm (in which cast the subcorpus argument must be given)}

\item{subcorpus}{if x is a single dtm, subcorpus should be a vector of the length and order of the dtm rows. Each value of subcorpus then represents a separate corpus.}

\item{method}{different ways to compare corpora. "each_to_all" compares each corpus to a corpos consisting of all other corpora. "chain" compares each corpus to a corpus consisting of the prevous [chain.lag] corpora. The order for chain comparison will be determined by sorting the dtm names (or the subcorpus values) with the order function.}

\item{return.df}{logical. If True, the results are returned as a data.frame. Otherwise as a list.}

\item{chain.lag}{An integer. If method is 'chain', this determines the number of previous corpora to which a corpus is compared. For example: if 3, then each corpus will be compared to a corpus consisting of the 3 previous corpora. Note that the first [chain.lag] corpora cannot be calculated, and will be ignored.}

\item{...}{additional arguments to be passed to the corpora.compare function}
}
\value{
a list or data.frame with the corpora.compare results for each dtm.
}
\description{
The corpora.compare function for a list of dtm's
}

