% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/corpus.r
\name{documents.compare}
\alias{documents.compare}
\title{Compare the documents in two corpora/dtms}
\usage{
documents.compare(dtm.x, dtm.y = NULL, measure = "cosine",
  min.similarity = 0.1, n.topsim = NULL)
}
\arguments{
\item{dtm.x}{the main document-term matrix}

\item{dtm.y}{the 'reference' document-term matrix. If NULL, documents of dtm.x are compared to each ohter}

\item{measure}{the measure that should be used to calculate similarity/distance/adjacency. Currently only cosine is supported}

\item{min.similarity}{a threshold for similarity. lower values are deleted. Set to 0.1 by default.}

\item{n.topsim}{An alternative or additional sort of threshold for similarity. Only keep the [n.topsim] highest similarity scores for x. Can return more than [n.topsim] similarity scores in the case of duplicate similarities.}
}
\value{
A data frame with rows corresponding to the terms in dtm and the statistics in the columns
}
\description{
Compare the documents in corpus dtm.x with reference corpus dtm.y.
}

