% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/corpus.r
\name{documents.window.compare}
\alias{documents.window.compare}
\title{Compare the documents in a dtm per time frame}
\usage{
documents.window.compare(dtm, document.date, window.size = 3,
  measure = "cosine", min.similarity = NULL, n.topsim = NULL,
  use.tfidf = F)
}
\arguments{
\item{dtm}{a document-term matrix in the tm format}

\item{document.date}{a vector of date class, of the same length and order as the documents (rows) of the dtm.}

\item{window.size}{the timeframe in days within which articles must occur in order to be compared. e.g., if 0, articles are only compared to articles of the same day. If 1, articles are compared to all articles of the previous, same or next day.}

\item{measure}{the measure that should be used to calculate similarity/distance/adjacency. Currently only cosine is supported}

\item{min.similarity}{a threshold for similarity. lower values are deleted}

\item{n.topsim}{An alternative or additional sort of threshold for similarity. Only keep the [n.topsim] highest similarities for x.}
}
\value{
A data frame with columns x, y and similarity.
}
\description{
Compare all documents within a document term matrix that are dated (e.g., pubished) within a given number of days (window.size) from each other.
}

