% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.r
\name{dtm.filter}
\alias{dtm.filter}
\title{Filter a dtm by selecting documents or terms}
\usage{
dtm.filter(dtm, documents = NULL, terms = NULL)
}
\arguments{
\item{dtm}{the dtm to filter}

\item{documents}{an optional vector of documents to include}

\item{terms}{an optional  vector of terms to include}
}
\value{
the filtered dtm
}
\description{
Filters the dtm and removes any empty rows or columns
}

