% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.r
\name{dtm.to.df}
\alias{dtm.to.df}
\title{Convert a dtm into a "triples" data frame}
\usage{
dtm.to.df(dtm, term_labels = colnames(dtm), doc_labels = rownames(dtm))
}
\arguments{
\item{dtm}{a document-term matrix}
}
\value{
a data frame with columns doc, term, and freq
}
\description{
Convert a dtm into a "triples" data frame
}

