% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/corpus.r
\name{dtm.to.dfm}
\alias{dtm.to.dfm}
\title{Convert a dtm into a quanteda dfm object (requires quanteda to be available)}
\usage{
dtm.to.dfm(dtm)
}
\arguments{
\item{dtm}{a document-term matrix}
}
\value{
a quanteda document-feature matrix
}
\description{
Convert a dtm into a quanteda dfm object (requires quanteda to be available)
}

