% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dtm.wordcloud}
\alias{dtm.wordcloud}
\title{Plot a word cloud from a dtm}
\usage{
dtm.wordcloud(dtm, nterms = 100, scale = c(6, 0.5), min.freq = 1,
  rot.per = 0.15, pal = brewer.pal(6, "YlGnBu"))
}
\arguments{
\item{dtm}{the document-term matrix}

\item{nterms}{the amount of words to plot (default 100)}

\item{scale}{the scale to plot (see wordcloud::wordcloud)}

\item{min.freq}{the minimum frquency to include (see wordcloud::wordcloud)}

\item{rot.per}{the percentage of vertical words (see wordcloud::wordcloud)}

\item{pal}{the colour model, see RColorBrewer}
}
\description{
Compute the term frequencies for the dtm and plot a word cloud with the top n topics
}

