% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.r, R/corpus.r
\name{dtmToSparseMatrix}
\alias{dtmToSparseMatrix}
\title{Transform a dtm into a sparse matrix.}
\usage{
dtmToSparseMatrix(dtm)

dtmToSparseMatrix(dtm)
}
\arguments{
\item{dtm}{a document-term matrix}

\item{dtm}{a document-term matrix}
}
\value{
a sparse matrix

a sparse matrix
}
\description{
Transform a dtm into a sparse matrix.

Transform a dtm into a sparse matrix.
}

