% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compare.r
\name{getOverlapTerms}
\alias{getOverlapTerms}
\title{Get the overlapping terms between two documents in a dtm}
\usage{
getOverlapTerms(x, y, dtm.x, dtm.y = NULL, similarity = NULL,
  termfreq = T)
}
\value{
a list or data.frame with the corpora.compare results for each dtm.
}
\description{
Get the overlapping terms between two documents in a dtm
}

