% Generated by roxygen2 (4.0.0): do not edit by hand
\name{lda.fit}
\alias{lda.fit}
\title{Estimate a topic model using the lda package}
\usage{
lda.fit(dtm, K = 50, num.iterations = 250, alpha = 50/K, eta = 0.01,
  burnin = 250, compute.log.likelihood = F)
}
\arguments{
\item{dtm}{a document term matrix (e.g. the output of \code{\link{amcat.dtm.create}})}

\item{K}{the number of clusters}

\item{num.iterations}{the number of iterations}

\item{alpha}{the alpha parameter}

\item{eta}{the eta parameter}
}
\value{
A fitted LDA model (see \code{\link{lda.collapsed.gibbs.sampler}})
}
\description{
Estimate an LDA topic model using the \code{\link{lda.collapsed.gibbs.sampler}} function
The parameters other than dtm are simply passed to the sampler but provide a workable default.
See the description of that function for more information
}

