% Generated by roxygen2 (4.0.0): do not edit by hand
\name{lda.match.meta}
\alias{lda.match.meta}
\title{Add document meta to LDA output}
\usage{
lda.match.meta(m, document.ids, meta, match.by = "id")
}
\arguments{
\item{m}{The output of \code{\link{lda.collapsed.gibbs.sampler}}}

\item{lda.document.ids}{A vector with document ids of the same length and order as the LDA output (matching the columns of m$document_sums)}

\item{meta}{A data.frame with document meta. Has to contain a vector to match the lda.document.ids}

\item{match.by}{The name of the vector in meta that matches the lda.document.ids}
}
\value{
The LDA output appended with document meta
}
\description{
Add a dataframe containing document meta to the output (a list) of \code{\link{lda.collapsed.gibbs.sampler}}.
}

