% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lda.plot.topic}
\alias{lda.plot.topic}
\title{Plots topic wordcloud, and attention over time and per category}
\usage{
lda.plot.topic(m, topic_nr, time_var, category_var, date_interval = "day",
  pct = F, value = "total")
}
\arguments{
\item{m}{The output of \code{\link{lda.collapsed.gibbs.sampler}}}

\item{The}{index of the topic (1 to K)}

\item{time_var}{A vector with time stamps (either numeric or Date class) of the same length and order of the documents (rows) in m$document_sums}

\item{category_var}{A vector with id values of the same length and order of the documents (rows) in m$document_sums}

\item{date_interval}{The interval for plotting the values over time. Can be: 'day', 'week', 'month' or 'year'}

\item{pct}{Show topic values as percentages}

\item{value}{Show topic values as 'total', or as 'relative' to the attention for other topics}
}
\value{
Nothing, just plots
}
\description{
Plots \code{\link{lda.plot.wordcloud}}, \code{\link{lda.plot.time}} and \code{\link{lda.plot.category}}
}

