% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lda.prepare.time.var}
\alias{lda.prepare.time.var}
\title{Change date object to date_interval}
\usage{
lda.prepare.time.var(time_var, date_interval)
}
\arguments{
\item{time_var}{A vector of Date values}

\item{date_interval}{The desired date_interval ('day','week','month', or 'year')}
}
\value{
A vector of Date values
}
\description{
Change date object to date_interval
}

