% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/corpus.r
\name{matchMatrixDims}
\alias{matchMatrixDims}
\title{Match the rows and columns of a list of matrices}
\usage{
matchMatrixDims(matrices)
}
\arguments{
\item{matrices}{a list of matrices}
}
\value{
a list of matrices
}
\description{
Resets the rows and columns of a list of matrices so that each matrix has the same dimensions.
}

