% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/corpus.r
\name{mergeDtmList}
\alias{mergeDtmList}
\title{Merge a list of dtm's}
\usage{
mergeDtmList(dtm_list)
}
\arguments{
\item{dtm_list}{a list of document-term matrices}
}
\value{
A single dtm
}
\description{
Merge a list of document term matrices, adding the attribute $subcorpus to indicate from which dtm a row originated.
}

