% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/corpus.r
\name{plotWords}
\alias{plotWords}
\title{Plot a wordcloud with words ordered and coloured according to a dimension (x)}
\usage{
plotWords(x, y = NULL, words = NULL, wordfreq = NULL, xlab = "",
  ylab = "", yaxt = "n", random.y = F, ...)
}
\value{
nothing
}
\description{
Plot a wordcloud with words ordered and coloured according to a dimension (x)
}

