% Generated by roxygen2 (4.0.0): do not edit by hand
\name{prepare.topics.plot.values}
\alias{prepare.topics.plot.values}
\title{Prepares the topic values per document for plotting}
\usage{
prepare.topics.plot.values(document_sums, break_var, topic_nr, pct = F,
  value = "total", filter = NULL)
}
\arguments{
\item{document_sums.}{A matrix where rows are topics and columns are articles. values represent how many words in the article belong to each topic. for collapsed.gibbs.sampler (lda package) get this matrix from the $document_sums object. for the topicmodels package, use the documentsums function (from corpustools).}

\item{break_var}{A break vector to aggregate topic values per document}

\item{The}{index of the topic (1 to K)}

\item{pct}{Show topic values as percentages}

\item{value}{Show topic values as 'total', or as 'relative' to the attention for other topics}
}
\value{
The aggregated/transformed topic values
}
\description{
Prepares the topic values per document for plotting
}

