% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/corpus.r
\name{setMatrixDims}
\alias{setMatrixDims}
\title{Set rows and columns of matrix to given rowdim and coldim vectors.}
\usage{
setMatrixDims(mat, rowdim, coldim)
}
\arguments{
\item{mat}{a matrix}

\item{rowdim}{a vector with rownames.}

\item{coldim}{a vector with colnames.}
}
\value{
a list of matrices
}
\description{
Set rows and columns of matrix to given rowdim and coldim vectors.
}

