% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{sotu}
\alias{sotu}
\title{Tokens from Bush and Obama's State of the Union addresses}
\format{sotu.dtm: A dataframe of tokens produced by using Stanford CoreNLP to lemmatize the State of the Union speeches; and sotu.meta the metadata for these documents}
\usage{
data(sotu)
}
\description{
Tokens from Bush and Obama's State of the Union addresses
}
\keyword{datasets}

