% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/corpus.r
\name{splitDtm}
\alias{splitDtm}
\title{Split a dtm into a list.}
\usage{
splitDtm(dtm, subcorpus)
}
\arguments{
\item{dtm}{a document-term matrix}

\item{subcorpus}{a vector that matches the documents (rows) of the dtm. The dtm will be split into separate dtms for each value of subcorpus.}
}
\value{
A list of dtms, named after the split_by values
}
\description{
Transform a dtm into a list containing the dtm in parts.
}

