% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/corpus.r
\name{term.statistics}
\alias{term.statistics}
\title{Compute some useful corpus statistics for a dtm}
\usage{
term.statistics(dtm)
}
\arguments{
\item{dtm}{a document term matrix (e.g. the output of \code{\link{dtm.create}})}
}
\value{
A data frame with rows corresponding to the terms in dtm and the statistics in the columns
}
\description{
Compute a number of useful statistics for filtering words: term frequency, idf, etc.
}

