% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/corpus.r
\name{term.time.statistics}
\alias{term.time.statistics}
\title{Compute corpus statistics for term use over time}
\usage{
term.time.statistics(dtm, document_date, time_interval = "day")
}
\arguments{
\item{dtm}{a document term matrix (e.g. the output of \code{\link{dtm.create}})}

\item{document_date}{a vector with dates that matches the rows in dtm}

\item{time_interval}{a character string indicating what time interval to use}
}
\value{
A data frame with rows corresponding to the terms in dtm and the statistics in the columns
}
\description{
Compute corpus statistics for term use over time. Calculates the spearman's rank order correlation of a term's use over time with a vector that gradually increases over time.
}

