% Generated by roxygen2 (4.0.1): do not edit by hand
\name{topics.plot.alltopics}
\alias{topics.plot.alltopics}
\title{Plot all topics}
\usage{
topics.plot.alltopics(document_sums, topics, time_var, category_var, path,
  date_interval = "day", value = "total")
}
\arguments{
\item{document_sums.}{A matrix where rows are topics and columns are articles. values represent how many words in the article belong to each topic. for collapsed.gibbs.sampler (lda package) get this matrix from the $document_sums object. for the topicmodels package, use the documentsums function (from corpustools).}

\item{topics}{a matrix in which rows are topics and columns are terms. values represent how prominent the word is in a topic. This matrix can for instance be obtained from the $topics object of the output of collapsed.gibbs.sampler, or when using the topicmodels package from posterior(m)$terms (where m is or stems from the LDA class)}

\item{time_var}{A vector with time stamps (either numeric or Date class) of the same length and order of the documents (rows) in m$document_sums}

\item{category_var}{A vector with id values of the same length and order of the documents (rows) in m$document_sums}

\item{path}{The path for a folder where output will be saved}

\item{date_interval}{The interval for plotting the values over time. Can be: 'day', 'week', 'month' or 'year'}
}
\value{
Nothing
}
\description{
Write plots for all topics with \code{\link{topics.plot.topic}} in designated folder
}

