% Generated by roxygen2 (4.0.0): do not edit by hand
\name{topics.plot.category}
\alias{topics.plot.category}
\title{Plots topic values per category}
\usage{
topics.plot.category(document_sums, topic_nr, category_var, pct = F,
  value = "total", return.values = F)
}
\arguments{
\item{document_sums.}{A matrix where rows are topics and columns are articles. values represent how many words in the article belong to each topic. for collapsed.gibbs.sampler (lda package) get this matrix from the $document_sums object. for the topicmodels package, use the documentsums function (from corpustools).}

\item{The}{index of the topic (1 to K)}

\item{category_var}{A vector with id values of the same length and order of the documents (rows) in m$document_sums}

\item{pct}{Show topic values as percentages}

\item{value}{Show topic values as 'total', or as 'relative' to the attention for other topics}

\item{return.values}{Logical. If true, data that is plotted is returned as a data.frame}
}
\value{
data.frame for plotted values
}
\description{
Plots the attention for a topic per category
}

