% Generated by roxygen2 (4.0.1): do not edit by hand
\name{topics.plot.time}
\alias{topics.plot.time}
\title{Plots topic values over time}
\usage{
topics.plot.time(document_sums, topic_nr, time_var, date_interval = "day",
  pct = F, value = "total", return.values = F)
}
\arguments{
\item{document_sums.}{A matrix where rows are topics and columns are articles. values represent how many words in the article belong to each topic. for collapsed.gibbs.sampler (lda package) get this matrix from the $document_sums object. for the topicmodels package, use the documentsums function (from corpustools).}

\item{The}{index of the topic (1 to K)}

\item{time_var}{A vector with time stamps (either numeric or Date class) of the same length and order of the documents (rows) in m$document_sums}

\item{date_interval}{The interval for plotting the values over time. Can be: 'day', 'week', 'month' or 'year'}

\item{pct}{Show topic values as percentages}

\item{value}{Show topic values as 'total', or as 'relative' to the attention for other topics}

\item{return.values}{Logical. If true, data that is plotted is returned as a data.frame}
}
\value{
data.frame for plotted values
}
\description{
Plots the attention for a topic over time
}

