% Generated by roxygen2 (4.0.0): do not edit by hand
\name{topics.plot.topic}
\alias{topics.plot.topic}
\title{Plots topic wordcloud, and attention over time and per category}
\usage{
topics.plot.topic(document_sums, topics, topic_nr, time_var, category_var,
  date_interval = "day", pct = F, value = "total")
}
\arguments{
\item{document_sums.}{A matrix where rows are topics and columns are articles. values represent how many words in the article belong to each topic. for collapsed.gibbs.sampler (lda package) get this matrix from the $document_sums object. for the topicmodels package, use the documentsums function (from corpustools).}

\item{topics}{a matrix in which rows are topics and columns are terms. values represent how prominent the word is in a topic. This matrix can for instance be obtained from the $topics object of the output of collapsed.gibbs.sampler, or when using the topicmodels package from posterior(m)$terms (where m is or stems from the LDA class)}

\item{The}{index of the topic (1 to K)}

\item{time_var}{A vector with time stamps (either numeric or Date class) of the same length and order of the documents (rows) in m$document_sums}

\item{category_var}{A vector with id values of the same length and order of the documents (rows) in m$document_sums}

\item{date_interval}{The interval for plotting the values over time. Can be: 'day', 'week', 'month' or 'year'}

\item{pct}{Show topic values as percentages}

\item{value}{Show topic values as 'total', or as 'relative' to the attention for other topics}
}
\value{
Nothing, just plots
}
\description{
Plots \code{\link{topics.plot.wordcloud}}, \code{\link{topics.plot.time}} and \code{\link{topics.plot.category}}
}

