% Generated by roxygen2 (4.0.0): do not edit by hand
\name{topmod.fill.time.gaps}
\alias{topmod.fill.time.gaps}
\title{Add empty values for pretty plotting}
\usage{
topmod.fill.time.gaps(d, date_interval)
}
\arguments{
\item{d}{A data.frame with the columns 'time' (Date) and 'value' (numeric)}

\item{date_interval}{The date_interval is required to know what the gaps are}
}
\value{
A data.frame with the columns 'time' (Date) and 'value' (numeric)
}
\description{
When plotting a timeline, gaps in date_intervals are ignored. For the attention for topics gaps should be considered as having value 0.
}

