% Generated by roxygen2 (4.0.0): do not edit by hand
\name{topmod.order.meta}
\alias{topmod.order.meta}
\title{Add document meta to LDA output}
\usage{
topmod.order.meta(m, meta, match.by = "id")
}
\arguments{
\item{m}{The output of \code{\link{LDA}}}

\item{meta}{A data.frame with document meta. Has to contain a vector to match the document.ids}

\item{match.by}{The name of the vector in meta that matches the document.ids}
}
\value{
The LDA output appended with document meta
}
\description{
Add a dataframe containing document meta to the output (a list) of \code{\link{lda.collapsed.gibbs.sampler}}.
}

