% Generated by roxygen2 (4.0.0): do not edit by hand
\name{topmod.plot.alltopics}
\alias{topmod.plot.alltopics}
\title{Plot all topics}
\usage{
topmod.plot.alltopics(m, time_var, category_var, path, date_interval = "day",
  value = "total")
}
\arguments{
\item{m}{The output of \code{\link{topmod.collapsed.gibbs.sampler}}}

\item{time_var}{A vector with time stamps (either numeric or Date class) of the same length and order of the documents (rows) in m$document_sums}

\item{category_var}{A vector with id values of the same length and order of the documents (rows) in m$document_sums}

\item{path}{The path for a folder where output will be saved}

\item{date_interval}{The interval for plotting the values over time. Can be: 'day', 'week', 'month' or 'year'}
}
\value{
Nothing
}
\description{
Write plots for all topics with \code{\link{topmod.plot.topic}} in designated folder
}

