% Generated by roxygen2 (4.0.1): do not edit by hand
\name{topmod.plot.category}
\alias{topmod.plot.category}
\title{Plots topic values per category}
\usage{
topmod.plot.category(m, topic_nr, category_var, pct = F, value = "total",
  return.values = F)
}
\arguments{
\item{m}{The output of \code{\link{LDA}}}

\item{The}{index of the topic (1 to K)}

\item{category_var}{A vector with id values of the same length and order of the documents (rows) in m$document_sums}

\item{pct}{Show topic values as percentages}

\item{value}{Show topic values as 'total', or as 'relative' to the attention for other topics}

\item{return.values}{Logical. If true, data that is plotted is returned as a data.frame}
}
\value{
data.frame for plotted values
}
\description{
Plots the attention for a topic per category
}

