% Generated by roxygen2 (4.0.0): do not edit by hand
\name{topmod.plot.wordcloud}
\alias{topmod.plot.wordcloud}
\title{Plot wordcloud for LDA topic}
\usage{
topmod.plot.wordcloud(m, topic_nr)
}
\arguments{
\item{m}{The output of \code{\link{LDA}}}

\item{topic_nr}{The index of the topic (1 to K)}
}
\value{
Nothing, just plots
}
\description{
Plots a wordcloud of the top words per topic
}

