% Generated by roxygen2 (4.0.0): do not edit by hand
\name{topmod.prepare.plot.values}
\alias{topmod.prepare.plot.values}
\title{Prepares the topic values per document for plotting}
\usage{
topmod.prepare.plot.values(m, break_var, topic_nr, pct = F, value = "total",
  filter = NULL)
}
\arguments{
\item{m}{The output of \code{\link{LDA}}}

\item{break_var}{A break vector to aggregate topic values per document}

\item{The}{index of the topic (1 to K)}

\item{pct}{Show topic values as percentages}

\item{value}{Show topic values as 'total', or as 'relative' to the attention for other topics}
}
\value{
The aggregated/transformed topic values
}
\description{
Prepares the topic values per document for plotting
}

