% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{wikinews_iraq}
\alias{wikinews_iraq}
\title{Document-term matrix based on wikinews articles on Iraq}
\format{iraq.dtm: A tm document-term matrix of wikinews articles, and iraq.meta: a data frame containing meta-information about the articles in the dtm}
\usage{
data(wikinews_iraq)
}
\description{
Document-term matrix based on wikinews articles on Iraq
}
\keyword{datasets}

