% Generated by roxygen2 (4.0.1): do not edit by hand
\name{chi2}
\alias{chi2}
\title{Compute the chi^2 statistic for a 2x2 crosstab containing the values
[[a, b], [c, d]]}
\usage{
chi2(a, b, c, d)
}
\description{
Compute the chi^2 statistic for a 2x2 crosstab containing the values
[[a, b], [c, d]]
}

