% Generated by roxygen2 (4.0.0): do not edit by hand
\name{dtm.create}
\alias{dtm.create}
\title{Create a document term matrix from a list of tokens}
\usage{
dtm.create(documents, terms, freqs)
}
\arguments{
\item{documents}{a vector of document names/ids}

\item{terms}{a vector of words of the same length as documents}

\item{freqs}{a vector of the frequency a a term in a document}
}
\value{
a document-term matrix  \code{\link{DocumentTermMatrix}}
}
\description{
Create a \code{\link{DocumentTermMatrix}} from a list of document ids, terms, and frequencies.
}

