% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/corpus.r
\name{dtmToSparseMatrix}
\alias{dtmToSparseMatrix}
\title{Transform a dtm into a sparse matrix.}
\usage{
dtmToSparseMatrix(dtm)
}
\arguments{
\item{dtm}{a document-term matrix}
}
\value{
a sparse matrix
}
\description{
Transform a dtm into a sparse matrix.
}

