% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lda.r
\name{lda.fit}
\alias{lda.fit}
\title{Estimate a topic model using the lda package}
\usage{
lda.fit(dtm, method = "Gibbs", K = 50, num.iterations = 500,
  alpha = 50/K, eta = 0.01, burnin = 250)
}
\arguments{
\item{dtm}{a document term matrix (e.g. the output of \code{\link{amcat.dtm.create}})}

\item{K}{the number of clusters}

\item{num.iterations}{the number of iterations}

\item{alpha}{the alpha parameter}

\item{eta}{the eta parameter}
}
\value{
A fitted LDA model (see \code{\link{LDA}})
}
\description{
Estimate an LDA topic model using the \code{\link{LDA}} function
The parameters other than dtm are simply passed to the sampler but provide a workable default.
See the description of that function for more information
}

