% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lda.r
\name{prepare.plot.values}
\alias{prepare.plot.values}
\title{Prepares the topic values per document for plotting}
\usage{
prepare.plot.values(m, break_var, topic_nr, pct = F, value = "relative",
  filter = NULL)
}
\arguments{
\item{m}{The output of \code{\link{LDA}}}

\item{break_var}{A break vector to aggregate topic values per document}

\item{pct}{Show topic values as percentages}

\item{value}{Show topic values as 'total', or as 'relative' to the attention for other topics}

\item{The}{index of the topic (1 to K)}
}
\value{
The aggregated/transformed topic values
}
\description{
Prepares the topic values per document for plotting
}

