% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lda.r
\name{topics.per.document}
\alias{topics.per.document}
\title{Get the topics per document, optionally merged with}
\usage{
topics.per.document(topics)
}
\arguments{
\item{dtm}{a document term matrix (e.g. the output of \code{\link{dtm.create}})}
}
\value{
A data frame with rows corresponding to the terms in dtm and the statistics in the columns
}
\description{
Return a data frame containing article metadata and topic occurence per document
}

